% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annualsummaryFunctions.R
\name{annual.summary.wq}
\alias{annual.summary.wq}
\title{summarize max and mean water quality conditions on an annual basis}
\usage{
annual.summary.wq(df, valuecol, datecol, locationcol)
}
\arguments{
\item{df}{data frame with estimated values, dates, location identifiers}

\item{valuecol}{string, name of column with water quality values}

\item{datecol}{string, name of column with dates (must be date format)}

\item{locationcol}{string, name of column with location identifiers}
}
\value{
dataframe of annual summaries
}
\description{
summarize max and mean water quality conditions on an annual basis
}
\examples{
data(estimatedrecord)
sumdata <- annual.summary.wq(df=estimatedrecord,valuecol="EstChlValue",datecol="ImageDate",
locationcol="StationID")
}
