% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorAnalysisFunctions.R
\name{climate.effect}
\alias{climate.effect}
\title{climate.effect}
\usage{
climate.effect(record, valuecol, climatevarcol, alternative = "two.sided")
}
\arguments{
\item{record}{dataframe with both record of water quality parameter and climate data (annual)}

\item{valuecol}{string, name of column with estimated or field-sampled water quality parameter}

\item{climatevarcol}{character, name of climate variable (column) of interest}

\item{alternative}{character string specifying alternative hypothesis ("two.sided","greater","less")}
}
\value{
results of wilcox test for differences in mean values (and, if overall, boxplots of water quality data)
}
\description{
Evaluates effects on HAB measures based on seasonal climate conditions
}
\examples{
data(utahlake_hab_climatedata)
effectresults <- climate.effect(record=utahlake_hab_climatedata,
valuecol="MaxAvgChl",climatevarcol="TotalWinterPrecip")
}
