% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorAnalysisFunctions.R
\name{climate.factor.compare}
\alias{climate.factor.compare}
\title{climate.factor.compare}
\usage{
climate.factor.compare(record, value, climaterecord, climatevar, lag = NULL,
  noevent, alternative = "two.sided", months = NULL)
}
\arguments{
\item{record}{dataframe with estimated historical record of water quality}

\item{value}{string, name of column with estimated or field-sampled water quality parameter}

\item{climaterecord}{dataframe with climate variables}

\item{climatevar}{character, name of climate variable (column) of interest}

\item{lag}{integer, number of days to lag the climate variable}

\item{noevent}{numeric, threshold for whether an event occured}

\item{alternative}{character string specifying alternative hypothesis ("two.sided","greater","less")}

\item{months}{an optional character string for if the t-test should be month specific}
}
\description{
Performs a t-test on water quality dataset for various climate variables
}
