% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorAnalysisFunctions.R
\name{climate_factor_station}
\alias{climate_factor_station}
\title{Climate Factor Impact}
\usage{
climate_factor_station(chlrecord, climaterecord, climatevar, lag = NULL,
  noevent, alternative = "two.sided", stations)
}
\arguments{
\item{chlrecord}{dataframe with estimated historical record of chl-a levels}

\item{climaterecord}{dataframe with climate variables}

\item{climatevar}{character, name of climate variable (column) of interest}

\item{lag}{integer, number of days to lag the climate variable}

\item{noevent}{numeric, threshold for whether an event occured}

\item{alternative}{character string specifying alternative hypothesis ("two.sided","greater","less")}

\item{stations}{list of unique stations to iterate over}
}
\description{
Performs a t-test on chl-a dataset for various climate variables
}
