% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{climatedata}
\alias{climatedata}
\title{Climate data}
\format{A data frame with 12238 rows and 3 variables:
\describe{
  \item{DATE}{date of observation}
  \item{PRCP}{precipitation volume (mm/day)}
  \item{TMAX}{max temperature (degrees C)}
  ...
}}
\source{
{NOAA Climate Data Archive}
}
\usage{
climatedata
}
\description{
A dataset containing precipitation and maximum daily temperature
for the Provo BYU NOAA Station
}
\keyword{datasets}
