% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.model.vars.R
\name{create.model.vars}
\alias{create.model.vars}
\title{create.model.vars}
\usage{
create.model.vars(filename, rowIndex = TRUE)
}
\arguments{
\item{filename}{CSV file with formatted surface reflectance data}

\item{rowIndex}{True or False, Indicates whether the first column in the formatted data file is a row index}
}
\value{
dataframe with variables used in model development or application
}
\description{
Create model variables used in model calibration and application
}
