% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.model.R
\name{cv.model}
\alias{cv.model}
\title{cv.model}
\usage{
cv.model(data, value, k, model, gof)
}
\arguments{
\item{data}{dataframe, limited to independent/dependent variables}

\item{value}{string, name of column with water quality values}

\item{k}{numeric, number of folds (will not be used if there are fewer observations than folds)}

\item{model}{formula}

\item{gof}{string, measure of the goodness of fit (PBIAS, R2, RMSE)}
}
\value{
list of training and testing goodness of fit
}
\description{
Use k-fold cross validation to evaluate the goodness of fit for a model
}
