% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{estimatedrecord}
\alias{estimatedrecord}
\title{Estimated  data}
\format{A data frame with 2041 rows and 5 variables:
\describe{
  \item{ImageDate}{date of image acquisition}
  \item{StationID}{location (corresponding to sampling locations)}
  \item{EstChlValue}{estimated chlorophyll value}
  \item{Lower}{lower end of confidence interval}
  \item{Upper}{upper end of confidence interval}
  ...
}}
\source{
{derived from models developed by
Carly Hansen and Landsat surface reflectance data}
}
\usage{
estimatedrecord
}
\description{
A dataset containing remotely sensed estimates of chlorophyll for Utah Lake
}
\keyword{datasets}
