% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorAnalysisFunctions.R
\name{lagpad}
\alias{lagpad}
\title{lagpad}
\usage{
lagpad(x, k)
}
\arguments{
\item{x}{climate data}

\item{k}{time step to lag (positive results in a forward shift, negative results in backwards shift)}
}
\description{
Lags a variable by a time step
}
