% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longterm_trend.R
\name{longterm_trend}
\alias{longterm_trend}
\title{Explore linear trends in the historic record}
\usage{
longterm_trend(record, by = "")
}
\arguments{
\item{record}{dataframe with record of chl-a levels. Must have columns: FieldValue, ImageDate.}

\item{by}{options for subsetting the data. month: to look at plots on a monthly basis; interval: plot the trend/confidence intervals for different lengths of the historic record}
}
\description{
Calculates linear trend and significance by month over the period of record
}
