% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modresults.R
\name{modresults}
\alias{modresults}
\title{modresults}
\usage{
modresults(model, data, value, title = "")
}
\arguments{
\item{model}{GLM model}

\item{data}{data for calibration/evaluation}

\item{value}{string, name of column with water quality values}

\item{title}{string, optional suffix for the title of the plot}
}
\value{
prints summary of model and plot of modeled vs. observed
}
\description{
Evaluate Model Performance (R2 and RMSE)
}
