% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TheilSenTrendFunctions.R
\name{monthlytrend.ts}
\alias{monthlytrend.ts}
\title{Explore long term monthly trends with Theil-Sen Estimator}
\usage{
monthlytrend.ts(record, value, date, months, var)
}
\arguments{
\item{record}{dataframe with record of estimated water quality}

\item{value}{string, name of column with values}

\item{date}{string, name of column with dates}

\item{months}{list of months}

\item{var}{string, aggregator (e.g. mean, max)}
}
\value{
summary of the Theil-Sen estimator
}
\description{
Calculates annual linear trend of average monthly values and significance of  with Theil-Sen Estimator (used for robust to non-normal data)
}
