% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{plot.record.errors}
\alias{plot.record.errors}
\title{plot.record.errors}
\usage{
\method{plot}{record.errors}(data, value, date, location,
  ylab = expression(paste("Chl-a (", mu, "g/L)")))
}
\arguments{
\item{data}{Dataframe with estimated values (value), dates (ImageDate), lake (Lake), lower and upper bounds (lower and upper)}

\item{value}{string, name of column with water quality values}

\item{date}{string, name of column with imagery dates}

\item{location}{string, name of column with location identifiers}

\item{ylab}{string, label for y axis}
}
\description{
Plots estimated record with error bars
}
