% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelApplicationFunctions.R
\name{plot_est_record.cal}
\alias{plot_est_record.cal}
\title{Plot record}
\usage{
plot_est_record.cal(histdata, caldata)
}
\arguments{
\item{histdata}{Dataframe with estimated values (FieldValue), dates (ImageDate), lake (Lake), lower and upper bounds (lower and upper)}

\item{caldata}{Dataframe with data used in Calibration (FieldValue, ImageDate, and Lake column)}
}
\description{
Plots estimated record with calibrated data
}
