% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{plotrecord}
\alias{plotrecord}
\title{plotrecord}
\usage{
plotrecord(data, datavalue, date, obsdata, obsdatavalue, obsdate, lake = "",
  labels = TRUE, ylab = expression(paste("Chl-a (", mu, "g/L)")))
}
\arguments{
\item{data}{Dataframe with estimated values}

\item{datavalue}{string, name of column with values in estimated dataframe}

\item{date}{string, name of column with date of imagery used for estimating values (must be date class)}

\item{obsdata}{Dataframe with Observed Data}

\item{obsdatavalue}{string, name of column with values in observed dataframe}

\item{obsdate}{string, name of column with date of observation (must be date class)}

\item{lake}{string, Name of Lake}

\item{labels}{optional for plotting}

\item{ylab}{string, label for y axis}
}
\value{
plot of estimated and observed data
}
\description{
Plots estimated and observed data
}
