% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{srdataforapplication}
\alias{srdataforapplication}
\title{Surface reflectance data from Landsat 5 and 7}
\format{A data frame with 2313 rows and 25 variables:
\describe{
  \item{Blue}{reflectance in the blue band}
  \item{Green}{reflectance in the green band}
  \item{ImageDate}{date of imagery acquisition}
  \item{NIR}{reflectance in the Near infrared band}
  \item{Red}{reflectance in the red band}
  \item{SWIR1}{reflectance in the short wave infrared 1 band}
  \item{SWIR2}{reflectance in the short wave infrared 2 band}
  \item{StationID}{location of sample}
  \item{Green_Blue}{reflectance in the green/blue band}
  \item{Red_Blue}{reflectance in the red/blue band}
  \item{Red_Green}{reflectance in the red/green band}
  \item{Red_NIR}{reflectance in the red/NIR band}
  \item{Red_SWIR1}{reflectance in the red/SWIR1 band}
  \item{Green_SWIR1}{reflectance in the green/SWIR1 band}
  \item{Blue_SWIR1}{reflectance in the blue/SWIR1 band}
  \item{Red_SWIR2}{reflectance in the red/SWIR2 band}
  \item{Green_SWIR2}{reflectance in the green/SWIR2 band}
  \item{Blue_SWIR2}{reflectance in the blue/SWIR2 band}
  \item{NIR_SWIR1}{reflectance in the nir/swir1 band}
  \item{NIR_SWIR2}{reflectance in the nir/swir2 band}
  \item{NIR_Blue}{reflectance in the nir/blue band}
  \item{NIR_Green}{reflectance in the nir/green band}
  \item{NDVI}{NDVI}
  \item{avgRGB}{average of reflectance in the visible bands}
  \item{avgSWIR}{average of reflectance in the SWIR bands}
  ...
}}
\source{
{Landsat}
}
\usage{
srdataforapplication
}
\description{
A dataset containing the surface reflectance for
locations in Utah Lake from Landsat 5 and 7
(precollection dataset) in Google Earth Engine, used for applying models.
}
\keyword{datasets}
