% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.model.R
\name{step.model}
\alias{step.model}
\title{step.model}
\usage{
step.model(data, imagerydate, value, modelvariables, timewindow, season,
  stepdirection, print.on = TRUE)
}
\arguments{
\item{data}{dataframe, formatted calibration data (model variables: field data and surface reflectance values)}

\item{imagerydate}{string, name of column with dates of imagery}

\item{value}{string, name of column with water quality values}

\item{modelvariables}{vector of strings with the names of columns for bands to consider}

\item{timewindow}{numeric, number of days to allow for near coincidence}

\item{season}{vector, months to include in model}

\item{stepdirection}{string, direction for stepwise regression ("backward","both","forward")}

\item{print.on}{boolean, option to print the resutls of the model (default is TRUE)}
}
\value{
list with the stepwise model and the modeled values
}
\description{
Use stepwise regression to parameterize model
}
