% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{utahsummermod}
\alias{utahsummermod}
\title{Example chlorophyll estimation model}
\format{A list of 30 items:
\describe{
  \item{coefficients}{}
  \item{residuals}{}
  \item{fitted.values}{}
  \item{effects}{}
  \item{R}{}
  \item{rank}{}
  \item{qr}{}
  \item{family}{}
  \item{linear.predictors}{}
  \item{deviance}{}
  \item{aic}{}
  \item{null.deviance}{}
  \item{iter}{}
  \item{weights}{}
  \item{prior.weights}{}
  \item{df.residual}{}
  \item{df.null}{}
  \item{y}{}
  \item{converged}{}
  \item{boundary}{}
  \item{model}{}
  \item{call}{}
  \item{formula}{}
  \item{terms}{}
  \item{data}{}
  \item{offset}{}
  \item{control}{}
  \item{method}{}
  \item{contrasts}{}
  \item{xlevels}{}
  ...
}}
\source{
{developed by Carly Hansen}
}
\usage{
utahsummermod
}
\description{
A dataset containing model information for Utah Lake - summer
}
\keyword{datasets}
