% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorAnalysisFunctions.R
\name{weather.effect}
\alias{weather.effect}
\title{weather.effect}
\usage{
weather.effect(wqrecord, imagedatecol, wqvarcol, climaterecord, climatedatecol,
  climatevarcol, maxlag = 0, noevent = NA, monthlist = NULL,
  locationcol = NULL)
}
\arguments{
\item{wqrecord}{dataframe with estimated historical record of water quality parameter}

\item{imagedatecol}{string, name of column with the date of the estimate (date of remotely sensed imagery)}

\item{wqvarcol}{string, name of column with estimated or field-sampled water quality parameter}

\item{climaterecord}{dataframe with climate variables}

\item{climatedatecol}{string, name of column with the date of the climate observation}

\item{climatevarcol}{character, name of climate variable (column) of interest}

\item{maxlag}{numeric, number of days to lag the climate effect}

\item{noevent}{numeric, threshold for whether an event occurred}

\item{monthlist}{vector, an optional vector containing the months to split the data on}

\item{locationcol}{string, name of column with unique location identifier, used if overall is FALSE}
}
\value{
results of kruskal wallis test for differences in mean values (and, if overall, boxplots of water quality data)
}
\description{
Evaluates difference in values based on weather conditions at quartiles and locations
}
\examples{
data(estimatedrecord)
data(climatedata)
effectresults <- weather.effect(wqrecord=estimatedrecord,imagedatecol="ImageDate",
wqvarcol="EstChlValue", climaterecord=climatedata,climatevarcol="TMAX",climatedatecol="DATE",
maxlag=7,noevent=0,monthlist=c("July","August"),locationcol="StationID")
}
