% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annualsummaryFunctions.R
\name{annual.summary.climate}
\alias{annual.summary.climate}
\title{summarize climate conditions on an annual basis}
\usage{
annual.summary.climate(data, date, value, parameter)
}
\arguments{
\item{data}{with estimated values, dates, location identifiers}

\item{date}{string, name of column with dates}

\item{value}{string, name of column with climate parameter values}

\item{parameter}{string, name of parameter ("Precipitation","Temperature")}
}
\description{
summarize climate conditions on an annual basis
}
