% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annualsummaryFunctions.R
\name{annual.summary.climate}
\alias{annual.summary.climate}
\title{summarize climate conditions on an annual basis}
\usage{
annual.summary.climate(df, datecol, valuecol, parameter, season)
}
\arguments{
\item{df}{with estimated values, dates, location identifiers}

\item{datecol}{string, name of column with dates}

\item{valuecol}{string, name of column with climate parameter values}

\item{parameter}{string, name of parameter ("Precipitation","Temperature")}

\item{season}{vector, names of months to include in analysis}
}
\value{
list of annual (seasonal) summaries
}
\description{
summarize climate conditions on an annual basis
}
\examples{
data(climatedata)
sumdata <- annual.summary.climate(df=climatedata,valuecol="PRCP",datecol="DATE",
parameter="Precipitation",season=c("January","February"))
}
