% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annualsummaryFunctions.R
\name{annual.summary.wq}
\alias{annual.summary.wq}
\title{summarize water quality conditions on an annual basis}
\usage{
annual.summary.wq(data, date, value, location)
}
\arguments{
\item{data}{with estimated values, dates, location identifiers}

\item{date}{string, name of column with dates}

\item{value}{string, name of column with water quality values}

\item{location}{string, name of column with location identifiers}
}
\description{
summarize water quality conditions on an annual basis
}
