% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.mod.seasonal.R
\name{apply.mod.seasonal}
\alias{apply.mod.seasonal}
\title{apply.mod.seasonal}
\usage{
apply.mod.seasonal(data, value, date, model, season, threshold)
}
\arguments{
\item{data}{dataframe with reflectance values}

\item{value}{string, name of column with water quality values}

\item{date}{string, name of column with imagery dates}

\item{model}{GLM (.Rdata)}

\item{season}{vector of months to include in the season}

\item{threshold}{numeric value at which to limit the allowed estimate)}
}
\description{
Apply GLM to remotely sensed record
}
