% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doy.max.trend.R
\name{doy.max.trend}
\alias{doy.max.trend}
\title{calculate trends in the occurrence of a maximum event}
\usage{
doy.max.trend(data, date, value, location)
}
\arguments{
\item{data}{with estimated values, dates, location identifiers}

\item{date}{string, name of column with dates}

\item{value}{string, name of column with water quality values}

\item{location}{string, name of column with location identifiers}
}
\description{
calculate trends in the occurrence of a maximum event
}
