% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatSRdata.R
\name{formatSRdata}
\alias{formatSRdata}
\title{formatSRdata}
\usage{
formatSRdata(data, imagerydate, samplingdate = "", location,
  datatype = "Estimated")
}
\arguments{
\item{data}{dataframe of surface reflectance data. Designed to work with reflectance values for bands from Landsat surface reflectance products (Blue, Red, Green, NIR, SWIR1, SWIR2, CloudMask) at specific point locations}

\item{imagerydate}{string, name of column for imagery dates}

\item{samplingdate}{string, name of column for sampling dates (only required if dataset is used for calibration)}

\item{location}{string, name of column for location identifier}

\item{datatype}{string, "Calibration" or "Estimated"}
}
\value{
dataframe with formatted data
}
\description{
Format surface reflectance data (used for application, not calibration)
}
