% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{plotrecord}
\alias{plotrecord}
\title{plotrecord}
\usage{
plotrecord(data, obsdata, lake, labels = TRUE,
  ylab = expression(paste("Chl-a (", mu, "g/L)")))
}
\arguments{
\item{data}{Dataframe with estimated values (value), dates (ImageDate), lake (Lake), lower and upper bounds (lower and upper)}

\item{obsdata}{Dataframe with Observed Data (Value, ImageDate)}

\item{lake}{string, Name of Lake}

\item{labels}{optional for plotting}

\item{ylab}{string, label for y axis}
}
\description{
Plots estimated and observed data
}
