% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{plotrecord}
\alias{plotrecord}
\title{plotrecord}
\usage{
plotrecord(data, datavalue, date, obsdata, obsdatavalue, obsdate,
  datacolors = c("grey", "red"), datashapes = c(1, 2), lake = "",
  title = "", ylab = expression(paste("Chl-a (", mu, "g/L)")))
}
\arguments{
\item{data}{Dataframe with estimated values}

\item{datavalue}{string, name of column with values in estimated dataframe}

\item{date}{string, name of column with date of imagery used for estimating values (must be date class)}

\item{obsdata}{Dataframe with Observed Data}

\item{obsdatavalue}{string, name of column with values in observed dataframe}

\item{obsdate}{string, name of column with date of observation (must be date class)}

\item{datacolors}{vector of length=2, specifying colors for estimated and observed datasets}

\item{datashapes}{vector of length=2, specifying shapes for estimated and observed datasets
#' @param lake string, Name of Lake}

\item{ylab}{string, label for y axis}

\item{labels}{optional for plotting}
}
\value{
plot of estimated and observed data
}
\description{
Plots estimated and observed data
}
